% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birth.R
\docType{data}
\name{air}
\alias{air}
\title{Air Transport Passengers Australia}
\format{
the format is: Annual Time-Series (1:27) from 1990 to 2016:
}
\source{
\pkg{fpp2}
}
\usage{
air
}
\description{
Total annual air passengers (in millions) including domestic and international
aircraft passengers of air carriers registered in Australia. 1970-2016.
}
\references{
Hyndman, R. & Khandakar, Y. (2008). Automatic time series forecasting: the
forecast package for \code{R}. \emph{Journal of Statistical Software}. 26(3),
1-22.\code{doi:	10.18637/jss.v027.i03}
}
\keyword{datasets}
