% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varstan.R
\name{as.stan}
\alias{as.stan}
\title{Convert to a stanfit object.}
\usage{
as.stan(object)
}
\arguments{
\item{object}{a varstan object.}
}
\value{
a stanfit object.
}
\description{
Convert a \code{varstan} object to a \code{stanfit} object of the
\pkg{rstan} package.
}
\examples{
\donttest{
 # Fitting a GARCH(1,1) model
 dat = garch(ipc,order = c(1,1,0))
 fit1 = varstan(dat,chains = 1)

 # Converting to a Stanfit object
 stanfit1 = as.stan(fit1)
}
}
\author{
Asael Alonzo Matamoros
}
