% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birth.R
\docType{data}
\name{aust}
\alias{aust}
\title{International Tourists to Australia: Total visitor nights.}
\format{
the format is: Quarterly Time-Series (1:44) from 1999 to 2015:
}
\source{
\pkg{fpp2}
}
\usage{
aust
}
\description{
Quarterly visitor nights (in millions) spent by international tourists
to Australia. 1999-2015
}
\references{
Hyndman, R. & Khandakar, Y. (2008). Automatic time series forecasting: the
forecast package for \code{R}. \emph{Journal of Statistical Software}. 26(3),
1-22.\code{doi:	10.18637/jss.v027.i03}
}
\keyword{datasets}
