% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{beta}
\alias{beta}
\title{Define a beta prior distribution}
\usage{
beta(shape1 = 2, shape2 = 2)
}
\arguments{
\item{shape1}{the first form parameter}

\item{shape2}{the second form parameter}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
beta(shape1,shape2)
}
\details{
Define a beta prior distribution using the hyper parameters
shape1 and shape2, by default a beta(2,2) distribution is
return.
}
