% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birth.R
\docType{data}
\name{birth}
\alias{birth}
\title{U.S. Monthly Live Births.}
\format{
the format is: Time-Series (1:373) from 1948 to 1979:
}
\source{
\pkg{astsa}
}
\usage{
birth
}
\description{
Monthly live births (adjusted) in thousands for the United States, 1948-1979.
}
\references{
http://www.stat.pitt.edu/stoffer/tsa4/
http://www.stat.pitt.edu/stoffer/tsda/
}
\keyword{datasets}
