% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{cauchy}
\alias{cauchy}
\title{Define a Cauchy prior distribution}
\usage{
cauchy(mu = 0, sd = 1)
}
\arguments{
\item{mu}{the location parameter mu}

\item{sd}{the standard desviation parameter sigma}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
cauchy(mu,sd)
}
\details{
Define a cauchy prior distribution using the hyper parameters
mu and sigma, by default a standard cauchy(0,1) distribution is
return.
}
