% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{chisq}
\alias{chisq}
\title{Define a chi square prior distribution}
\usage{
chisq(df = 7)
}
\arguments{
\item{df}{the degree freedom parameter df}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
chisq(df)
}
\details{
Define a gamma prior distribution using the degree freedom df hyper parameter,
by default an chisq(7) distribution is return.
}
