% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birth.R
\docType{data}
\name{demgbp}
\alias{demgbp}
\title{DEM/GBP exchange rate log-returns}
\format{
the format is: Time-Series (1:350) from 1984 to 1985:
}
\source{
\pkg{bayesGARCH}
}
\usage{
demgbp
}
\description{
The vector dem2gbp contains daily observations of the Deutschmark vs British
Pound foreign exchange rate log-returns. This data set has been promoted as
an informal benchmark for GARCH time-series software validation. See
McCullough and Renfro (1999), and Brooks, Burke, and Persand (2001) for details.
The nominal returns are expressed in percent as in Bollerslev and Ghysels (1996).
The sample period is from January 3, 1984, to December 31, 1991, for a total
of 1974 observations.
}
\references{
Engle, R. (1982). Autoregressive Conditional Heteroscedasticity with Estimates of
the Variance of United Kingdom Inflation. \emph{Econometrica}, 50(4), 987-1007.
\code{url: http://www.jstor.org/stable/1912773}.

Bollerslev, T. (1986). Generalized autoregressive conditional heteroskedasticity.
\emph{Journal of Econometrics}. 31(3), 307-327.
\code{doi: https://doi.org/10.1016/0304-4076(86)90063-1}.

Ardia, D. and Hoogerheide, L. (2010). Bayesian Estimation of the GARCH(1,1) Model
with Student-t Innovations. \emph{The R Journal}. 2(7), 41-47.
\code{doi: 10.32614/RJ-2010-014}.
}
\keyword{datasets}
