% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{exponential}
\alias{exponential}
\title{Define an exponential prior distribution}
\usage{
exponential(rate = 1)
}
\arguments{
\item{rate}{the rate parameter lambda in exponential distribution}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
exponential(rate)
}
\details{
Define a gamma prior distribution using the rate hyper parameter,
by default an exponential(1) distribution is return.
}
