% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{fitted.varstan}
\alias{fitted.varstan}
\title{Expected Values of the Posterior Predictive Distribution}
\usage{
\method{fitted}{varstan}(object, robust = FALSE, ...)
}
\arguments{
\item{object}{a \code{varstan} object.}

\item{robust}{a bool value, if its \code{TRUE} it returns the median of the
posterior distribution, and if its \code{FALSE} it returns the mean, by default
is the \code{FALSE} value.}

\item{...}{Further arguments passed to  \code{posterior_predict}.}
}
\value{
A time series \code{(ts)} of predicted \emph{mean} response values.
}
\description{
The function returns the posterior estimate of the fitted values
of a \code{varstan} model, similar to the fit_values functions of other
packages.
}
\details{
This function returns a time series of the predicted \emph{mean} response values.
}
\seealso{
\code{posterior_predict.varstan}
}
\author{
Asael Alonzo Matamoros
}
