% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{ggnorm}
\alias{ggnorm}
\title{\code{qqplot} with normal \code{qqline}}
\usage{
ggnorm(y, title = NULL, add.normal = TRUE)
}
\arguments{
\item{y}{a numeric vector or an object of the \code{ts} class containing a stationary time series.}

\item{title}{a string with the plot's title.}

\item{add.normal}{Add a normal density function for comparison.}
}
\value{
None.
}
\description{
Plot the quantile-quantile plot and quantile-quantile line using ggplot.
}
\examples{
x = rnorm(100)
ggnorm(x)

}
\author{
Asael Alonzo Matamoros
}
