% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{ggpacf}
\alias{ggpacf}
\title{\code{pacf} plot.}
\usage{
ggpacf(y, title = NULL)
}
\arguments{
\item{y}{a numeric vector or an object of the \code{ts} class containing a
stationary time series.}

\item{title}{a string with the plot's title.}
}
\value{
None.
}
\description{
Plot of the partial autocorrelation function for a univariate time series.
}
\examples{
x = rnorm(100)
ggpacf(x)

}
\author{
Mitchell O'Hara-Wild and Asael Alonzo Matamoros
}
