% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{inverse.gamma}
\alias{inverse.gamma}
\title{Define an inverse gamma prior distribution}
\usage{
inverse.gamma(shape = 2, rate = 1)
}
\arguments{
\item{shape}{the form parameter alpha in gamma distribution}

\item{rate}{the rate parameter beta in gamma distribution}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
inverse.gamma(shape,rate)
}
\details{
Define a inverse.gamma prior distribution using the hyper parameters
shape and rate, by default an inverse.gamma(2,1) distribution is
return.

If sigma has a gamma distribution then 1/sigma has n  inverse gamma
distribution. The rate parameter is the inverse of an scale parameter.
}
