% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc.R
\docType{data}
\name{ipc}
\alias{ipc}
\title{Monthly inflation coefficients from 1980-2018.}
\format{
A time series of monthly data from 1980 to 2018.
}
\source{
https://www.bch.hn/series_estadisticas.php
}
\usage{
ipc
}
\description{
Monthly return coefficients for the inflation. An economic indicator of
a country's economy.
}
\keyword{datasets}
