% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{jeffrey}
\alias{jeffrey}
\title{Define a non informative Jeffrey's prior for the degree freedom hyper parameter}
\usage{
jeffrey()
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
jeffey.df()
}
\details{
Define a non informative Jeffrey's prior distribution, by default an jeffrey.df( )
distribution is return.

This prior can only be used in garch models with t-student innovations, or Bekk
models with generalized t-student distribution.
}
