% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{laplace}
\alias{laplace}
\title{Define a Laplace prior distribution}
\usage{
laplace(mu = 0, sd = 1)
}
\arguments{
\item{mu}{the location parameter mu}

\item{sd}{the standard deviation parameter sigma}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
laplace(mu,sd)
}
\details{
Define a Laplace prior distribution using the hyper parameters
mu and sigma, by default a standard Laplace distribution is
return.

The laplace distribution is exactly the same as the double exponential distribution
}
