% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{LKJ}
\alias{LKJ}
\title{Define a LKJ matrix prior distribution}
\usage{
LKJ(df = 2)
}
\arguments{
\item{df}{the degree freedom parameter df}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
LKJ(df)
}
\details{
Define a Lewandowski Kurowicka and Joe (LKJ) matrix correlation prior
distribution using the degree freedom df hyper parameter,by default
a LKJ(2) distribution is return.
}
\references{
Lewandowski D, Kurowicka D, Joe H (2009). "Generating random correlation matrices based
on vines and extended onion method." Journal of Multivariate Analysis, 100(9), 1989 2001.
ISSN 0047-259X. doi:https://doi.org/10.1016/j.jmva.2009.04.008.
URL: http://www.sciencedirect.com/science/article/pii/S0047259X09000876.
}
