% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{normal}
\alias{normal}
\title{Define a normal prior distribution}
\usage{
normal(mu = 0, sd = 1)
}
\arguments{
\item{mu}{the location parameter mu}

\item{sd}{the standard desviation parameter sigma}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
normal(mu,sd)
}
\details{
Define a normal prior distribution using the hyper parameters
mu and sigma, by default a standard normal distribution is
return.
}
