% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birth.R
\docType{data}
\name{oildata}
\alias{oildata}
\title{Annual oil production in Saudi Arabia}
\format{
the format is: Annual Time-Series (1:18) from 1996 to 2013:
}
\source{
\pkg{fpp2}
}
\usage{
oildata
}
\description{
Annual oil production (millions of tonnes), Saudi Arabia, 1965-2013.
}
\references{
Hyndman, R. & Khandakar, Y. (2008). Automatic time series forecasting: the
forecast package for \code{R}. \emph{Journal of Statistical Software}. 26(3),
1-22.\code{doi:	10.18637/jss.v027.i03}
}
\keyword{datasets}
