% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.garch}
\alias{print.garch}
\title{Print a \code{garch} model}
\usage{
\method{print}{garch}(x, ...)
}
\arguments{
\item{x}{a \code{garch} model.}

\item{...}{additional values need in print methods.}
}
\value{
None. prints the object
}
\description{
Print a \code{garch} model
}
