% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Holt}
\alias{print.Holt}
\title{Print a Holt model}
\usage{
\method{print}{Holt}(x, ...)
}
\arguments{
\item{x}{a Holt model from the varstan package}

\item{...}{additional values need in print methods}
}
\value{
None. prints the object.
}
\description{
Print a Holt model
}
