% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Hw}
\alias{print.Hw}
\title{Print a Holt-Winter model}
\usage{
\method{print}{Hw}(x, ...)
}
\arguments{
\item{x}{a Hw from the varstan package}

\item{...}{additional values need in print methods}
}
\value{
None. prints the object.
}
\description{
Print a Holt-Winter model
}
