% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.LocalLevel}
\alias{print.LocalLevel}
\title{Print a Local Level model}
\usage{
\method{print}{LocalLevel}(x, ...)
}
\arguments{
\item{x}{a LocalLevel model from the varstan package}

\item{...}{additional values need in print methods}
}
\value{
None. prints the object.
}
\description{
Print a Local Level model
}
