% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Sarima}
\alias{print.Sarima}
\title{Print a \code{Sarima} model.}
\usage{
\method{print}{Sarima}(x, ...)
}
\arguments{
\item{x}{a \code{Sarima} model.}

\item{...}{additional values need in print methods.}
}
\value{
None. prints the object
}
\description{
Print a \code{Sarima} model.
}
