% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ssm}
\alias{print.ssm}
\title{Print a state-space model.}
\usage{
\method{print}{ssm}(x, ...)
}
\arguments{
\item{x}{a \code{ssm} model.}

\item{...}{additional values need in print methods.}
}
\value{
None. prints the object.
}
\description{
Print a state-space model.
}
