% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SVM}
\alias{print.SVM}
\title{Print a Stochastic Volatility model}
\usage{
\method{print}{SVM}(x, ...)
}
\arguments{
\item{x}{a SVM model from the varstan package}

\item{...}{additional values need in print methods}
}
\value{
None. prints the object
}
\description{
Print a Stochastic Volatility model
}
