% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.varstan}
\alias{print.varstan}
\title{Print a varstan object}
\usage{
\method{print}{varstan}(x, ...)
}
\arguments{
\item{x}{a varstan object}

\item{...}{additional values need in print methods}
}
\value{
None. prints the object.
}
\description{
Print a varstan object
}
