% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_error.R
\name{residuals.varstan}
\alias{residuals.varstan}
\title{Generic function and method for extract the residual of a \code{varstan} object}
\usage{
\method{residuals}{varstan}(object, robust = FALSE, ...)
}
\arguments{
\item{object}{a \code{varstan} object.}

\item{robust}{a bool value, if its \code{TRUE} it returns the median of the
posterior distribution, and if its \code{FALSE} it returns the mean. By default,
is the \code{FALSE} value.}

\item{...}{Further arguments passed to  \code{posterior_residual}.}
}
\value{
An array with the posterior estimate of the residuals of the time series model,
}
\description{
The function returns the posterior estimate of the residuals
of a \code{varstan} model, similar to the residual functions of other
packages.
}
\details{
This function only extracts the point-wise estimate of the time series residuals
for extracting all the data use \code{extract_stan()} or posterior_intervals
function
}
\author{
Asael Alonzo Matamoros
}
