% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{uniform}
\alias{uniform}
\title{Define a uniform prior distribution}
\usage{
uniform(min = 0, max = 1)
}
\arguments{
\item{min}{the first form parameter}

\item{max}{the second form parameter}
}
\value{
a numerical vector interpreted as a prior in Stan
}
\description{
uniform(shape1,shape2)
}
\details{
Define a beta prior distribution using the hyper parameters
min and max, by default a uniform(0,1) distribution is
return.
}
