% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{check_residuals}
\alias{check_residuals}
\title{Visual check of residuals in a \code{varstan} object.}
\usage{
check_residuals(object, ...)
}
\arguments{
\item{object}{a varstan object.}

\item{...}{Other plotting parameters to affect the plot.}
}
\value{
A graph object from ggplot2.
}
\description{
Performs a visual check of residuals in time series models, this method is inspired in
the \code{check.residuals} function provided by the \code{forecast} package.
}
\examples{
\donttest{
 library(astsa)
 sf1 = auto.sarima(ts = birth)
 # fitted model
 check_resiudals(sf1)
}

}
\author{
Asael Alonzo Matamoros.
}
