% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.garch}
\alias{print.garch}
\title{Print a garch model}
\usage{
\method{print}{garch}(x, ...)
}
\arguments{
\item{x}{a garch model from the varstan package}

\item{...}{additional values need in print methods}
}
\value{
None. prints the object
}
\description{
Print a garch model
}
