% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.naive}
\alias{print.naive}
\title{Print a \code{naive} model}
\usage{
\method{print}{naive}(x, ...)
}
\arguments{
\item{x}{a \code{naive} model.}

\item{...}{additional values need in print methods.}
}
\value{
None. prints the object
}
\description{
Print a \code{naive} model
}
