% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.naive}
\alias{print.naive}
\title{Print a naive model}
\usage{
\method{print}{naive}(x, ...)
}
\arguments{
\item{x}{a navie model from the varstan package}

\item{...}{additional values need in print methods}
}
\description{
Print a naive model
}
