% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Sarima}
\alias{print.Sarima}
\title{Print a Sarima model}
\usage{
\method{print}{Sarima}(x, ...)
}
\arguments{
\item{x}{a Sarima model from the varstan package}

\item{...}{additional values need in print methods}
}
\value{
None. prints the object
}
\description{
Print a Sarima model
}
